Update usage of SimaiFile in tests, improve formatting
Updated the usage of SimaiFile in tests for properly using the IDisposable interface. SimaiFile instances are now wrapped with a using statement which ensures that any system resources the class holds are properly disposed of when they are no longer needed. This change is intended to prevent possible file-handle/memory leaks.

Simultaneously simplified the creation of a StreamWriter in SimaiFile's constructor by removing unnecessary 'using' and made tests formatting more consistent for better readability.