Refactor note and slide reading to simplify timing override
This commit changes the handling of tempo override in SlideReader and NoteReader. Instead of locally creating a new timing override in SlideReader, we now create it at the calling site (in NoteReader) and pass it to SlideReader. This change makes the behaviour more explicit, easier to reason about, and reduces unnecessary object creation.