Refactor Deserializer.cs for better efficiency
The changes aim to improve processing efficiency by reordering the operations in Deserializer.cs. Previously, we would return _chart early if currentNoteCollection was null, and then check and add the currentNoteCollection to noteCollections. Now we're setting _chart.FinishTiming before these processes. This is to ensure every noteCollection gets added to the list before returning _chart, even if currentNoteCollection is null. It is a neater and more efficient structure as unused objects are cleaned up sooner.